extract getActionButton to its own component
